src/requirements.txt
Flask
pymongo
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
